BC-1893 #comment tinggal remove tenant's folder di 'model'-nya

+++ commons/src/main/java/org/soluvas/commons/shell/TenantRemoveCommand.java
@@ -4,6 +4,7 @@ import javax.annotation.Nullable;
 
 import org.apache.felix.gogo.commands.Argument;
 import org.apache.felix.gogo.commands.Command;
+import org.apache.felix.gogo.commands.Option;
 import org.soluvas.commons.tenant.ProvisionData;
 import org.soluvas.commons.tenant.TenantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,14 +22,19 @@ import com.google.common.collect.ImmutableSet;
 @Command(scope="tenant", name="remove", description="Removes tenant(s) using TenantRepository#remove(java.util.Set).")
 public class TenantRemoveCommand extends ExtCommandSupport {
 	
+	
 	/**
 	 * Not required so it doesn't break apps that don't use {@link TenantRepository}. 
 	 */
 	@Autowired(required=false) @Nullable
 	private TenantRepository<ProvisionData> tenantRepo;
 	
+	@Option(name="--forge", aliases="-f", description="Forge for removing..")
+	private boolean f;
+	
 	@Argument(name="tenantId ...", required=true, multiValued=true, description="Tenant ID(s)")
 	private String[] tenantIds;
+	
 	
 	public TenantRemoveCommand() {
 		super(false);
@@ -36,8 +42,10 @@ public class TenantRemoveCommand extends ExtCommandSupport {
 	
 	@Override
 	protected Void doExecute() throws Exception {
-		Preconditions.checkNotNull(tenantRepo, "TenantRepository bean must present");
-		tenantRepo.remove(ImmutableSet.copyOf(tenantIds));
+		if (f) {
+			Preconditions.checkNotNull(tenantRepo, "TenantRepository bean must present");
+			tenantRepo.remove(ImmutableSet.copyOf(tenantIds));
+		}
 		return null;
 	}
